requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.3

# ─── AI / Claude ──────────────────────────────────────────────────────────────
anthropic==0.28.0
openai==1.30.1
transformers==4.41.2          # FinBERT
torch==2.3.1                  # FinBERT inference
sentence-transformers==3.0.0

# ─── NLP ──────────────────────────────────────────────────────────────────────
spacy==3.7.4
# run: python -m spacy download en_core_web_lg

# ─── News / Social ────────────────────────────────────────────────────────────
newsapi-python==0.2.7
feedparser==6.0.11
tweepy==4.14.0
praw==7.7.1
aiohttp==3.9.5
httpx==0.27.0

# ─── Technical Analysis ───────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
# TA-Lib requires system install: brew install ta-lib
TA-Lib==0.4.29
scipy==1.13.0

# ─── Market Data ──────────────────────────────────────────────────────────────
alpha-vantage==2.3.1
yfinance==0.2.40
pandas-datareader==0.10.0
fredapi==0.5.2

# ─── ZeroMQ Bridge ────────────────────────────────────────────────────────────
pyzmq==26.0.3

# ─── Database ─────────────────────────────────────────────────────────────────
supabase==2.4.5
redis==5.0.5
asyncpg==0.29.0

# ─── API Server ───────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
websockets==12.0
python-jose[cryptography]==3.3.0

# ─── Scheduling ───────────────────────────────────────────────────────────────
APScheduler==3.10.4

# ─── Backtesting ──────────────────────────────────────────────────────────────
backtesting==0.3.3

# ─── Utilities ────────────────────────────────────────────────────────────────
loguru==0.7.2
rich==13.7.1
sentry-sdk==2.5.1
arrow==1.3.0
tenacity==8.3.0              # retry logic
